Begin splitting up files
