+++ src/lib.rs
@@ -13,177 +13,18 @@ use diesel::{
         methods::{ExecuteDsl, LimitDsl, LoadQuery},
         RunQueryDsl,
     },
-    r2d2::{self, ManageConnection, R2D2Connection},
+    r2d2::R2D2Connection,
     QueryResult,
 };
-use std::sync::{Arc, Mutex};
-use thiserror::Error;
 use tokio::task;
 
-/// A connection manager which implements [`bb8::ManageConnection`] to
-/// integrate with bb8.
-///
-/// ```no_run
-/// use async_bb8_diesel::AsyncRunQueryDsl;
-/// use diesel::prelude::*;
-/// use diesel::pg::PgConnection;
-///
-/// table! {
-///     users (id) {
-///         id -> Integer,
-///     }
-/// }
-///
-/// #[tokio::main]
-/// async fn main() {
-///     use users::dsl;
-///
-///     // Creates a Diesel-specific connection manager for bb8.
-///     let mgr = async_bb8_diesel::DieselConnectionManager::<PgConnection>::new("localhost:1234");
-///     let pool = bb8::Pool::builder().build(mgr).await.unwrap();
-///
-///     // You can acquire connections to the pool manually...
-///     diesel::insert_into(dsl::users)
-///         .values(dsl::id.eq(1337))
-///         .execute_async(&*pool.get().await.unwrap())
-///         .await
-///         .unwrap();
-///
-///     // ... Or just issue them to the pool directly.
-///     diesel::insert_into(dsl::users)
-///         .values(dsl::id.eq(1337))
-///         .execute_async(&pool)
-///         .await
-///         .unwrap();
-/// }
-/// ```
-#[derive(Clone)]
-pub struct DieselConnectionManager<T> {
-    inner: Arc<Mutex<r2d2::ConnectionManager<T>>>,
-}
+mod diesel_connection;
+mod diesel_connection_manager;
+mod error;
 
-impl<T: Send + 'static> DieselConnectionManager<T> {
-    pub fn new<S: Into<String>>(database_url: S) -> Self {
-        Self {
-            inner: Arc::new(Mutex::new(r2d2::ConnectionManager::new(database_url))),
-        }
-    }
-
-    async fn run_blocking<R, F>(&self, f: F) -> R
-    where
-        R: Send + 'static,
-        F: Send + 'static + FnOnce(&r2d2::ConnectionManager<T>) -> R,
-    {
-        let cloned = self.inner.clone();
-        tokio::task::spawn_blocking(move || f(&*cloned.lock().unwrap()))
-            .await
-            // Intentionally panic if the inner closure panics.
-            .unwrap()
-    }
-}
-
-#[async_trait]
-impl<T> bb8::ManageConnection for DieselConnectionManager<T>
-where
-    T: R2D2Connection + Send + 'static,
-{
-    type Connection = DieselConnection<T>;
-    type Error = ConnectionError;
-
-    async fn connect(&self) -> Result<Self::Connection, Self::Error> {
-        self.run_blocking(|m| m.connect())
-            .await
-            .map(DieselConnection::new)
-            .map_err(ConnectionError::Checkout)
-    }
-
-    async fn is_valid(
-        &self,
-        conn: &mut bb8::PooledConnection<'_, Self>,
-    ) -> Result<(), Self::Error> {
-        let c = DieselConnection(conn.0.clone());
-        self.run_blocking(move |m| {
-            m.is_valid(&mut *c.inner())?;
-            Ok(())
-        })
-        .await
-    }
-
-    fn has_broken(&self, _: &mut Self::Connection) -> bool {
-        // Diesel returns this value internally. We have no way of calling the
-        // inner method without blocking as this method is not async, but `bb8`
-        // indicates that this method is not mandatory.
-        false
-    }
-}
-
-/// An async-safe analogue of any connection that implements
-/// [`diesel::Connection`].
-///
-/// These connections are created by [`DieselConnectionManager`].
-///
-/// All blocking methods within this type delegate to
-/// [`tokio::task::spawn_blocking`], meaning they won't block
-/// any asynchronous work or threads.
-pub struct DieselConnection<C>(pub(crate) Arc<Mutex<C>>);
-
-impl<C> DieselConnection<C> {
-    pub fn new(c: C) -> Self {
-        Self(Arc::new(Mutex::new(c)))
-    }
-
-    // Accesses the underlying connection.
-    //
-    // As this is a blocking mutex, it's recommended to avoid invoking
-    // this function from an asynchronous context.
-    fn inner(&self) -> std::sync::MutexGuard<'_, C> {
-        self.0.lock().unwrap()
-    }
-}
-
-/// Syntactic sugar around a Result returning an [`ConnectionError`].
-pub type ConnectionResult<R> = Result<R, ConnectionError>;
-
-/// Errors returned directly from DieselConnection.
-#[derive(Error, Debug)]
-pub enum ConnectionError {
-    #[error("Failed to checkout a connection: {0}")]
-    Checkout(#[from] diesel::r2d2::Error),
-
-    #[error("Failed to issue a query: {0}")]
-    Query(#[from] diesel::result::Error),
-}
-
-/// Syntactic sugar around a Result returning an [`PoolError`].
-pub type PoolResult<R> = Result<R, PoolError>;
-
-/// Describes an error performing an operation from a connection pool.
-///
-/// This is a superset of [`ConnectionError`] which also may
-/// propagate errors attempting to access the connection pool.
-#[derive(Error, Debug)]
-pub enum PoolError {
-    #[error("Failed to checkout a connection: {0}")]
-    Connection(#[from] ConnectionError),
-
-    #[error("BB8 Timeout accessing connection")]
-    Timeout,
-}
-
-impl From<diesel::result::Error> for PoolError {
-    fn from(error: diesel::result::Error) -> Self {
-        PoolError::Connection(ConnectionError::Query(error))
-    }
-}
-
-impl From<bb8::RunError<ConnectionError>> for PoolError {
-    fn from(error: bb8::RunError<ConnectionError>) -> Self {
-        match error {
-            bb8::RunError::User(e) => PoolError::Connection(e),
-            bb8::RunError::TimedOut => PoolError::Timeout,
-        }
-    }
-}
+pub use diesel_connection::DieselConnection;
+pub use diesel_connection_manager::DieselConnectionManager;
+pub use error::{ConnectionError, ConnectionResult, PoolError, PoolResult};
 
 /// An async variant of [`diesel::connection::SimpleConnection`].
 #[async_trait]
